+++ galilei/emulator.py
@@ -1,43 +1,6 @@
-from functools import partial
-
 import numpy as np
 import torch
-from sklearn.gaussian_process import GaussianProcessRegressor, kernels
-from sklearn.metrics import mean_squared_error
 from sklearn.model_selection import train_test_split
-from torch import nn, optim
-from torch.nn import functional as F
-from torch.utils.data import DataLoader, TensorDataset
-from tqdm import tqdm
-
-
-class Net(nn.Module):
-    def __init__(self, idim, odim, hidden=[64, 64], activation=F.relu):
-        """
-        Simple fully connected neural network with ReLU activation
-
-        Parameters
-        ----------
-        idim : int
-            Input dimension
-        odim : int
-            Output dimension
-        hidden : list of int
-            List of hidden layer sizes
-        """
-        super(Net, self).__init__()
-        self.activation = activation
-        self.layers = nn.ModuleList()
-        self.layers.append(nn.Linear(idim, hidden[0]))
-        for i in range(1, len(hidden)):
-            self.layers.append(nn.Linear(hidden[i - 1], hidden[i]))
-        self.layers.append(nn.Linear(hidden[-1], odim))
-
-    def forward(self, out):
-        for layer in self.layers[:-1]:
-            out = self.activation(layer(out))
-        out = self.layers[-1](out)
-        return out
 
 
 class Emulator:
@@ -308,400 +271,6 @@ class Emulator:
         self._load(store)
 
 
-class TorchEmulator(Emulator):
-    def __init__(
-        self,
-        NNClass=Net,
-        epochs=500,
-        optimizer=None,
-        criterion=None,
-        lr=0.01,
-        weight_decay=0,
-        momentum=0.9,
-        NN_kwargs={},
-        scheduler=None,
-        batch_size=32,
-        device=None,
-        **kwargs,
-    ):
-        """
-        A class used to emulate data using a neural network based on the PyTorch library.
-
-        Parameters
-        ----------
-        NNClass : nn.Module, optional
-            The neural network class to use. Default is Net.
-        epochs : int, optional
-            The number of epochs to train the neural network. Default is 500.
-        optimizer : torch.optim.Optimizer, optional
-            The optimizer to use for training the neural network. If None, uses Stochastic Gradient Descent (SGD) optimizer with default parameters.
-        criterion : torch.nn.modules.loss._Loss, optional
-            The loss function to use for training the neural network. If None, uses Mean Squared Error (MSE) loss function.
-        lr : float, optional
-            The learning rate for the optimizer. Default is 0.01.
-        weight_decay : float, optional
-            The weight decay (L2 penalty) for the optimizer. Default is 0.
-        momentum : float, optional
-            The momentum factor for the optimizer. Default is 0.9.
-        NN_kwargs : dict, optional
-            The keyword arguments to pass to the neural network class constructor. Default is an empty dictionary.
-        **kwargs
-            Additional keyword arguments to pass to the base Emulator class constructor.
-
-        Attributes
-        ----------
-        NNClass : callable
-            A callable that returns an instance of the neural network class with the given keyword arguments.
-        epochs : int
-            The number of epochs to train the neural network.
-        optimizer : torch.optim.Optimizer
-            The optimizer to use for training the neural network.
-        criterion : torch.nn.modules.loss._Loss
-            The loss function to use for training the neural network.
-        """
-        super(TorchEmulator, self).__init__(**kwargs)
-        self.NNClass = partial(NNClass, **NN_kwargs)
-        self.epochs = epochs
-        self.optimizer = (
-            optimizer
-            if optimizer is not None
-            else partial(optim.SGD, lr=lr, momentum=momentum, weight_decay=weight_decay)
-        )
-        self.criterion = criterion if criterion is not None else nn.MSELoss()
-        self.scheduler = scheduler
-        self.batch_size = batch_size
-        if device is None:
-            self.device = "cuda" if torch.cuda.is_available() else "cpu"
-        else:
-            self.device = device
-
-    def _train(self, X_train, Y_train):
-        """
-        Train the emulator
-
-        Parameters
-        ----------
-        X_train : np.ndarray
-            Training input
-        Y_train : np.ndarray
-            Training output
-        """
-        device = self.device
-        # convert the data to PyTorch tensors and send to GPU if available
-        X_train = torch.as_tensor(X_train, dtype=torch.float32, device=device)
-        Y_train = torch.as_tensor(Y_train, dtype=torch.float32, device=device)
-
-        # create TensorDatasets and DataLoaders
-        train_data = TensorDataset(X_train, Y_train)
-        train_loader = DataLoader(train_data, batch_size=32, shuffle=True)
-
-        # build model, optimizer and loss function
-        model = self.NNClass(X_train.shape[-1], Y_train.shape[-1]).to(device)
-        optimizer = self.optimizer(model.parameters())
-        scheduler = self.scheduler(optimizer) if self.scheduler is not None else None
-
-        # train the model
-        print("Training emulator...")
-        pbar = tqdm(range(self.epochs))
-        for epoch in pbar:
-            running_loss = 0.0
-            for i, (inputs, Y) in enumerate(train_loader):
-                # zero the parameter gradients
-                optimizer.zero_grad()
-
-                # forward + backward + optimize
-                outputs = model(inputs)
-                loss = self.criterion(outputs, Y)
-                loss.backward()
-                optimizer.step()
-
-                # print statistics
-                running_loss += loss.item()
-            pbar.set_postfix({"loss": f"{running_loss/(i+1):.5f}"})
-            if scheduler is not None:
-                scheduler.step()
-
-        self.model = model
-
-    def _test(self, X_test, Y_test):
-        """
-        Test the emulator on a test set
-
-        Parameters
-        ----------
-        X_test : array-like
-            The test set input data
-        Y_test : array-like
-            The test set output data
-        """
-        device = self.device
-        X_test = torch.as_tensor(X_test, dtype=torch.float32, device=device)
-        Y_test = torch.as_tensor(Y_test, dtype=torch.float32, device=device)
-        # Test the model
-        if len(X_test) > 0:
-            with torch.no_grad():
-                test_loss = 0.0
-                for i, data in enumerate(zip(X_test, Y_test)):
-                    inputs, Y = data
-                    outputs = self.model(inputs)
-                    loss = self.criterion(outputs, Y)
-                    test_loss += loss.item()
-                avg_test_loss = test_loss / (i + 1)
-            print("Ave Test loss: {:.3f}".format(avg_test_loss))
-
-    def _predict(self, x):
-        """
-        Predict the output of the emulator for a given input
-
-        Parameters
-        ----------
-        x : array-like
-            The input data
-
-        Returns
-        -------
-        np.ndarray
-            The predicted output
-        """
-        x = torch.as_tensor(x, dtype=torch.float32, device=self.device)
-        res = self.model(x).detach().cpu().numpy()
-        return res
-
-    def _save(self, store):
-        """
-        Save the emulator model to a file
-
-        Parameters
-        ----------
-        store: dict
-            The dictionary that will be stored to disk
-
-        """
-        store["model"] = self.model.state_dict()
-
-    def _load(self, store):
-        """
-        Load the emulator model from a file
-
-        Parameters
-        ----------
-        store: dict
-            The dictionary that will be stored to disk
-        """
-        self.model = self.NNClass(store["in"], store["out"])
-        self.model.load_state_dict(store["model"])
-        self.model.to(self.device)
-
-
-class SklearnEmulator(Emulator):
-    def __init__(
-        self,
-        model=None,
-        kernel_kwargs={},
-        model_kwargs={"n_restarts_optimizer": 10},
-        **kwargs,
-    ):
-        """
-        A class used to emulate data using a machine learning model based on the scikit-learn library.
-
-        Parameters
-        ----------
-        model : sklearn.base.TransformerMixin, optional
-            The machine learning model to use. If None, uses a Gaussian Process model with an RBF kernel from scikit-learn with default parameters.
-        kernel_kwargs : dict, optional
-            The keyword arguments to pass to the RBF kernel constructor if a Gaussian Process model is used. Default is an empty dictionary.
-        model_kwargs : dict, optional
-            The keyword arguments to pass to the machine learning model constructor if one is provided. Default sets n_restarts_optimizer to 5.
-        **kwargs
-            Additional keyword arguments to pass to the base Emulator class constructor.
-
-        Attributes
-        ----------
-        model : sklearn.base.TransformerMixin
-            The machine learning model used for emulation.
-        """
-
-        super(SklearnEmulator, self).__init__(**kwargs)
-        if model is None:
-            # use a gaussian process model with RKB kernel from sklearn if none is provided
-            model = GaussianProcessRegressor(
-                kernel=kernels.RBF(**kernel_kwargs), **model_kwargs
-            )
-        self.model = model
-
-    def _train(self, X_train, Y_train):
-        """
-        Train the emulator
-
-        Parameters
-        ----------
-        X_train : np.ndarray
-            Training input
-        Y_train : np.ndarray
-            Training output
-        """
-        self.model.fit(X_train, Y_train)
-
-    def _test(self, X_test, Y_test):
-        """
-        Test the emulator on a test set
-
-        Parameters
-        ----------
-        X_test : array-like
-            The test set input data
-        Y_test : array-like
-            The test set output data
-        """
-        if len(X_test) > 0:
-            test_loss = mean_squared_error(Y_test, self.model.predict(X_test))
-            print("Ave Test loss: {:.3f}".format(test_loss))
-
-    def _predict(self, x):
-        """
-        Predict the output of the emulator for a given input
-
-        Parameters
-        ----------
-        x : array-like
-            The input data
-
-        Returns
-        -------
-        np.ndarray
-            The predicted output
-        """
-        res = self.model.predict([x])[0]
-        return res
-
-    def _save(self, store):
-        """
-        Save the emulator model to a file
-
-        Parameters
-        ----------
-        store: dict
-            The dictionary that will be stored to disk
-
-        """
-        store["model"] = self.model
-
-    def _load(self, store):
-        """
-        Load the emulator model from a file
-
-        Parameters
-        ----------
-        store: dict
-            The dictionary that will be stored to disk
-        """
-        self.model = store["model"]
-
-
-class GPyEmulator(Emulator):
-    def __init__(
-        self,
-        kernel_kwargs={},
-        model_kwargs={},
-        **kwargs,
-    ):
-        """
-        A class used to emulate data using a machine learning model based on the GPy library.
-
-        Parameters
-        ----------
-        kernel_kwargs : dict, optional
-            The keyword arguments to pass to the RBF kernel constructor if a Gaussian Process model is used. Default is an empty dictionary.
-        model_kwargs : dict, optional
-            The keyword arguments to pass to the machine learning model constructor if one is provided. Default is an empty dictionary.
-        **kwargs
-            Additional keyword arguments to pass to the base Emulator class constructor.
-
-        Attributes
-        ----------
-        model : GPy.models.GPRegression
-            The machine learning model used for emulation.
-        """
-        super(GPyEmulator, self).__init__(**kwargs)
-        self.kernel_kwargs = kernel_kwargs
-        self.model_kwargs = model_kwargs
-        self.model = None
-
-    def _train(self, X_train, Y_train):
-        """
-        Train the emulator
-
-        Parameters
-        ----------
-        X_train : np.ndarray
-            Training input
-        Y_train : np.ndarray
-            Training output
-        """
-        import GPy
-
-        kernel = GPy.kern.RBF(input_dim=X_train.shape[-1], **self.kernel_kwargs)
-        model = GPy.models.GPRegression(X_train, Y_train, kernel, **self.model_kwargs)
-        model.optimize()
-        self.model = model
-
-    def _test(self, X_test, Y_test):
-        """
-        Test the emulator on a test set
-
-        Parameters
-        ----------
-        X_test : array-like
-            The test set input data
-        Y_test : array-like
-            The test set output data
-        """
-        if len(X_test) > 0:
-            test_loss = np.mean((Y_test - self.model.predict(X_test)[0]) ** 2)
-            print("Ave Test loss: {:.5f}".format(test_loss))
-
-    def _predict(self, x):
-        """
-        Predict the output of the emulator for a given input
-
-        Parameters
-        ----------
-        x : array-like
-            The input data
-
-        Returns
-        -------
-        np.ndarray
-            The predicted output
-        """
-        res = self.model.predict(np.array([x]))
-        return res[0][0]
-
-    def _save(self, store):
-        """
-        Save the emulator model to a file
-
-        Parameters
-        ----------
-        store: dict
-            The dictionary that will be stored to disk
-        """
-        store["model"] = self.model.to_dict()
-
-    def _load(self, store):
-        """
-        Load the emulator model from a file
-
-        Parameters
-        ----------
-        store: dict
-            The dictionary that will be stored to disk
-        """
-        import GPy
-
-        self.model = GPy.models.GPRegression.from_dict(store["model"])
-
-
 # the main emulator interface
 class emulate:
     def __init__(
@@ -758,12 +327,20 @@ class emulate:
 
         # select the backend
         if backend == "torch":
+            try:
+                from .backends.torch import TorchEmulator
+            except ImportError:
+                raise ImportError(
+                    "PyTorch is not installed. Please install it to use the PyTorch backend."
+                )
             self.emulator_class = TorchEmulator
         elif backend == "sklearn":
+            from .backends.sklearn import SklearnEmulator  # installed by default
+
             self.emulator_class = SklearnEmulator
         elif backend == "gpy":
             try:
-                import GPy
+                from .backends.GPy import GPyEmulator
             except ImportError:
                 raise ImportError(
                     "GPy is not installed. Please install it to use the GPy backend."
